lib: Introduce netlink-devlink library.
The devlink interface was introduced [0] in the Linux 4.6 time
frame and has since gained traction among multiple hardware
vendors.

The devlink-port [1] and devlink-info[2] interfaces are
particularly useful for managing NICs connected to multiple
distinct CPUs such as SmartNICs.

In such a topology it would be useful to be able to offload
Open vSwitch and OVN onto the NIC SoC operating system and this
library will help with discovering and managing ports representing
resources made available on the host from the NIC SoC side.

This has been proposed for inclusion in Open vSwitch [3], but let's
carry it here until it lands.

0: https://lore.kernel.org/netdev/1456504351-18871-1-git-send-email-jiri@resnulli.us/
1: https://github.com/torvalds/linux/blob/master/Documentation/networking/devlink/devlink-port.rst
2: https://github.com/torvalds/linux/blob/master/Documentation/networking/devlink/devlink-info.rst
3: http://patchwork.ozlabs.org/project/openvswitch/patch/20210506102840.285473-1-frode.nordahl@canonical.com/

Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>