representor: Prepare data store for runtime updates.
The current data store is suitable for the lookups needed to meet
CMS requests.

However it is insufficient to meet the needs of performing runtime
maintenance of the cached data.

To reliably update the data we will monitor updates from devlink
as representors are created or removed.  The devlink data provides
us with information about association between PF and VF
representors as well as ifindex and name of the netdevs.

Since we are not dumping all the data for each iteration we do
need to rely on our stored information bout the PF in order to
update the PF MAC+VF number hmap.  To deal with this part
we add a hmap that maps from bus_name+dev_name to port nodes
containing information about PHYSICAL and PF representor ports.

On modern systems network interfaces are subject to renaming so
monitoring devlink alone will unfortunately not provide us with
all the information we need as the netdev name will be out of date
by the time we get around to use it.  We also do not want to make
multiple system calls on each iteration to lookup netdev name from
ifindex.

To deal with this we will monitor udev events in addition to
devlink so that we can update our lookup tables if a `move` event
arrives for one of the interfaces in our cache.  To deal with
this part we add a hmap using ifindex as hash.

Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>