representor store: Add source and netdev_renamed attributes.
On modern Linux systems netdev names are subject to renaming.
To avoid unnecessary churn as a result of runtime updates to
interfaces in our lookup tables we add attributes to track source
of data as well as a flag to tell whether the netdev rename has
occurred or not.

These attributes will allow us to assume that the rename has
occurred when dumping the data on startup, and expect a move
event when updating data as a result of runtime updates when
that is appropriate according to system configuration.

This is not a perfect solution as there is a corner case where
the ovn-controller might be started or restarted while the
system is in the middle of adding netdevs, but to be pragmatic
this will most likely not be the case very often.

Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>