vif-plug-representor: Log reason for failure during udev initialization.
In the case of failure during udev initialization, the reason is
currently not logged.  This makes it difficult to triage any
issues.

The udev functions returning pointers will populate errno on
failure and the udev functions with an int return will pass on
the errno value as a negative integer.

Reported-at: https://launchpad.net/bugs/2034700
Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>