vif-plug-representor: Lower severity of set udev recv buf size failure.
Setting the udev receive buffer size is a privileged operation,
and the ovn-controller may not always be allowed to set it.

Before this patch this message would be logged when running unit
tests as a non-root user and/or when the ovn-controller is ran in
an unprivilieged container.

The OVS/OVN unit tests often check and fail on any unexpected
WARN, ERR or EMER level messages.

Lower severity to INFO.

Reported-at: https://launchpad.net/bugs/2034700
Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>